parallel: drop --keep-order to stream test output with -j
The Tap 14 specification explicitly allows for out-of-order Test Points,
which means we do not need to use --keep-order. The primary benefit of
doing this is that users will get more feedback during long test runs as
to whether tests are passing or not.

> Test Points *may* be output in any order, but any Test Point ID provided
> must be within the range described by the Plan.

Technically, Tap 13 didn't disallow out-of-order Test Points (and implies
they are allowed) but at the time of implementing "bats -j" support this
wasn't clear enough to justify the risk of producing spec-invalid output.

Fixes: da65d1bcecfe ("bats: add support for parallel job execution with GNU parallel")
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>